more refinement
